# FIX: Update topic/post counter correctly when category has zero topics