@davidtaylorhq 2019-12-20
This does work... but having read it again I think we'll be better off splitting this into multiple queries. This one huge query is very difficult to understand.

 I will take another look.