@davidtaylorhq 2019-12-20
Ok, much much simpler now. Ready for review