@davidtaylorhq 2019-12-20
I checked the performance locally, and it was identical. Of course, with a bigger database the story is slightly different. With my 'simpler' method, it takes about 40% longer on a massive database. I've gone back to the 'hard to read, but working' single query, and there is no noticeable performance impact. I guess here we have to sacrifice readability for performance.

On meta: (average over 5 runs)
  - Original (broken) method: 7.5s
  - New 'simpler' method: 9s
  - 'Complicated' query method: 7.5s

On a larger site with 500k topics (1 run):
  - Old method: 83.1s
  - New 'simpler' method: 116.7s
  - 'Complicated' query method: 77.8